requirements.txt
fastapi
uvicorn
python-multipart
pydantic
faiss-cpu
sentence-transformers
openai
PyPDF2
numpy
python-dotenv